Changed file set up to match UML, Split up Product and review into their own folders, and now reviews are their own object meaning it will be easier to add sutomer reviews and append them to a product. API call working to get products and to search products by keyword.
