+++ app/models/Product.py
@@ -1,28 +1,7 @@
 from datetime import datetime
 from typing import List, Dict, Optional
+from app.models.Review import Review
 import json
-
-class Review:
-    def __init__(self, review_id: str, user_id: str, user_name: str, rating: float, title: str, content: str):
-        self.review_id = review_id
-        self.user_id = user_id
-        self.user_name = user_name
-        self.rating = rating
-        self.title = title
-        self.content = content
-        self.date_posted = datetime.now().isoformat()
-
-    def to_dict(self) -> Dict:
-        return {
-            "review_id": self.review_id,
-            "user_id": self.user_id,
-            "user_name": self.user_name,
-            "rating": self.rating,
-            "review_title": self.title,
-            "review_content": self.content,
-            "date_posted": self.date_posted
-        }
-
 
 class Product:
     def __init__(self, product_id: str, product_name: str, category: str, discounted_price: str, actual_price: str, discount_percentage: str, rating: float, rating_count: int, about_product: str, img_link: str, product_link: str):
@@ -37,11 +16,12 @@ class Product:
         self.about_product = about_product
         self.img_link = img_link
         self.product_link = product_link
-        self.reviews: List[Review] = []
+        self.reviews: List[Review] = [] #composition (product has maany reviews)
 
     def add_review(self, review: Review):
         self.reviews.append(review)
 
+    #converts product and reviews to dict for json saving
     def to_dict(self) -> Dict:
         return {
             "product_id": self.product_id,
