+++ app/repositories/IProductRepository.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+class IProductRepository(ABC):
+    @abstractmethod
+    def load_products(self):
+        pass
+
+    @abstractmethod
+    def save_products(self, products):
+        pass
