+++ test/test_products.py
@@ -1,8 +1,8 @@
-from app.repositories.ProductRepo import ProductRepository
+from app.repositories.FileProductRepository import FileProductRepository
 
 
 def test_load_products():
     # tests loading products from JSON
-    repo = ProductRepository("app/data/Products.json")
+    repo = FileProductRepository("app/data/Products.json")
     products = repo.load_products()
     assert len(products) > 0
